requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
httpx==0.27.2
python-dotenv==1.0.1
aiosqlite==0.20.0
jinja2==3.1.4
aio-pika==9.4.3
